FAP update. Learning about "auto"
auto in a range based-for-loop defaults to returning by reference.

Several inventory pieces of code have been improved to reflect this better understanding of C++.

The custom functions to interface with FAP have been made more robust and more general to prevent unwanted repetition.

+++ Fight_MovementManager.cpp
@@ -4,7 +4,7 @@
 # include "Source\Fight_MovementManager.h"
 # include <random> // C++ base random is low quality.
 
-#define DISTANCE_METRIC CUNYAIModule::getProperSpeed(unit) * 8;
+#define DISTANCE_METRIC (int)CUNYAIModule::getProperSpeed(unit) * 8;
 
 using namespace BWAPI;
 using namespace Filter;
